Use sbt-tpolecat for scalac options
... and apply these options on itself.
Compile `core` in strict mode now
and fix extra compile errors that came out of it.

+++ core/src/main/scala/precog/domain/PullRequestDraft.scala
@@ -43,7 +43,7 @@ final case class PullRequestDraft(
 object PullRequestDraft {
   implicit class ToPullRequestSyntax(pr: PullRequestDraft) {
     def toPullRequest: PullRequest = PullRequest(
-      id = pr.id,
+      id = pr.id.toLong,
       number = pr.number,
       state = pr.state,
       title = pr.title,
